requirements.txt
# Flight Controller
Adafruit-Blinka
adafruit-circuitpython-bmp280
adafruit-circuitpython-mpu6050
RPi.GPIO

# Dashboard
flask
